DeliaOptions.bulkInsertEnabled
